Move gamepad files to their own folder
