+++ components/script/dom/gamepad/gamepad.rs
@@ -8,7 +8,10 @@ use dom_struct::dom_struct;
 use embedder_traits::{GamepadSupportedHapticEffects, GamepadUpdateType};
 use js::typedarray::{Float64, Float64Array};
 
-use super::bindings::buffer_source::HeapBufferSource;
+use super::gamepadbuttonlist::GamepadButtonList;
+use super::gamepadhapticactuator::GamepadHapticActuator;
+use super::gamepadpose::GamepadPose;
+use crate::dom::bindings::buffer_source::HeapBufferSource;
 use crate::dom::bindings::codegen::Bindings::GamepadBinding::{GamepadHand, GamepadMethods};
 use crate::dom::bindings::codegen::Bindings::GamepadButtonListBinding::GamepadButtonListMethods;
 use crate::dom::bindings::inheritance::Castable;
@@ -18,10 +21,7 @@ use crate::dom::bindings::root::{Dom, DomRoot};
 use crate::dom::bindings::str::DOMString;
 use crate::dom::event::Event;
 use crate::dom::eventtarget::EventTarget;
-use crate::dom::gamepadbuttonlist::GamepadButtonList;
 use crate::dom::gamepadevent::{GamepadEvent, GamepadEventType};
-use crate::dom::gamepadhapticactuator::GamepadHapticActuator;
-use crate::dom::gamepadpose::GamepadPose;
 use crate::dom::globalscope::GlobalScope;
 use crate::dom::window::Window;
 use crate::script_runtime::{CanGc, JSContext};
