+++ components/script/dom/gamepad/gamepadhapticactuator.rs
