+++ components/script/dom/gamepad/gamepadpose.rs
@@ -5,7 +5,7 @@
 use dom_struct::dom_struct;
 use js::typedarray::{Float32, Float32Array};
 
-use super::bindings::buffer_source::HeapBufferSource;
+use crate::dom::bindings::buffer_source::HeapBufferSource;
 use crate::dom::bindings::codegen::Bindings::GamepadPoseBinding::GamepadPoseMethods;
 use crate::dom::bindings::reflector::{Reflector, reflect_dom_object};
 use crate::dom::bindings::root::DomRoot;
